refactor(server): extract the reinstall wait loop so it can be shared
